refactor: Magic넘버처리 및 Console 파일 view패키지로 이동
